# backend/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/cars": {
      "get": {
        "operationId": "CarsController_getAllCars",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number to retrieve (1-indexed)",
            "schema": {
              "minimum": 1,
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Maximum number of items to return per page",
            "schema": {
              "minimum": 1,
              "default": 10,
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "brandId",
            "required": false,
            "in": "query",
            "description": "Filter results to a single brand by exact brand identifier",
            "schema": {
              "example": "brand-1",
              "type": "string"
            }
          },
          {
            "name": "modelId",
            "required": false,
            "in": "query",
            "description": "Filter results to a single model by exact model identifier. Typically combined with brandId.",
            "schema": {
              "example": "model-1",
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Field used to sort the vehicle list. Only list-visible fields are exposed here.",
            "schema": {
              "example": "brand",
              "enum": [
                "brand",
                "model",
                "total"
              ],
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "description": "Sort direction applied to sortBy",
            "schema": {
              "default": "asc",
              "example": "asc",
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Vehicle catalog returned successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/PaginatedResponseDto"
                    },
                    {
                      "properties": {
                        "items": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/CarSummary"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "access_token": []
          }
        ],
        "summary": "List vehicles with pagination, filters and sorting",
        "tags": [
          "Vehicles"
        ]
      },
      "post": {
        "operationId": "CarsController_createCar",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCarDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Vehicle created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Car"
                }
              }
            }
          },
          "400": {
            "description": "Invalid vehicle payload"
          },
          "403": {
            "description": "Forbidden"
          },
          "409": {
            "description": "Vehicle conflicts with an existing license plate or brand/model combination"
          }
        },
        "security": [
          {
            "access_token": []
          }
        ],
        "summary": "Create a new vehicle entry (ADMIN only)",
        "tags": [
          "Vehicles"
        ]
      }
    },
    "/cars/export/excel": {
      "get": {
        "operationId": "CarsController_exportCarsToExcel",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number to retrieve (1-indexed)",
            "schema": {
              "minimum": 1,
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Maximum number of items to return per page",
            "schema": {
              "minimum": 1,
              "default": 10,
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "brandId",
            "required": false,
            "in": "query",
            "description": "Filter results to a single brand by exact brand identifier",
            "schema": {
              "example": "brand-1",
              "type": "string"
            }
          },
          {
            "name": "modelId",
            "required": false,
            "in": "query",
            "description": "Filter results to a single model by exact model identifier. Typically combined with brandId.",
            "schema": {
              "example": "model-1",
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Field used to sort the vehicle list. Only list-visible fields are exposed here.",
            "schema": {
              "example": "brand",
              "enum": [
                "brand",
                "model",
                "total"
              ],
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "description": "Sort direction applied to sortBy",
            "schema": {
              "default": "asc",
              "example": "asc",
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Excel file generated successfully"
          }
        },
        "security": [
          {
            "access_token": []
          }
        ],
        "summary": "Export the filtered vehicle table to a real Excel file",
        "tags": [
          "Vehicles"
        ]
      }
    },
    "/cars/{id}": {
      "get": {
        "operationId": "CarsController_getCarById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Vehicle identifier",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Vehicle returned successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Car"
                }
              }
            }
          },
          "404": {
            "description": "Vehicle not found"
          }
        },
        "security": [
          {
            "access_token": []
          }
        ],
        "summary": "Retrieve a vehicle by identifier",
        "tags": [
          "Vehicles"
        ]
      },
      "put": {
        "operationId": "CarsController_updateCar",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Vehicle identifier",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCarDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Vehicle updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Car"
                }
              }
            }
          },
          "400": {
            "description": "Invalid vehicle payload. PUT expects the full CreateCarDto contract."
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Vehicle not found"
          },
          "409": {
            "description": "Vehicle conflicts with an existing license plate or brand/model combination"
          }
        },
        "security": [
          {
            "access_token": []
          }
        ],
        "summary": "Update an existing vehicle entry (ADMIN only)",
        "tags": [
          "Vehicles"
        ]
      },
      "delete": {
        "operationId": "CarsController_deleteCar",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Vehicle identifier",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Vehicle deleted successfully"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Vehicle not found"
          }
        },
        "security": [
          {
            "access_token": []
          }
        ],
        "summary": "Delete a vehicle entry (ADMIN only)",
        "tags": [
          "Vehicles"
        ]
      }
    },
    "/cars/{id}/document": {
      "get": {
        "operationId": "CarsController_getCarDocumentMetadata",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Vehicle identifier",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Document metadata returned successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadedCarDocumentResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Vehicle or document not found"
          }
        },
        "security": [
          {
            "access_token": []
          }
        ],
        "summary": "Retrieve metadata for the single document linked to a vehicle",
        "tags": [
          "Vehicles"
        ]
      },
      "delete": {
        "operationId": "CarsController_deleteCarDocument",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Vehicle identifier",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Document deleted successfully"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Vehicle or document not found"
          }
        },
        "security": [
          {
            "access_token": []
          }
        ],
        "summary": "Delete the single document linked to a vehicle (ADMIN only)",
        "tags": [
          "Vehicles"
        ]
      },
      "post": {
        "operationId": "CarsController_uploadCarDocument",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Vehicle identifier",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "example": "Ficha tecnica ITV"
                  },
                  "documentType": {
                    "type": "string",
                    "enum": [
                      "invoice",
                      "inspection",
                      "insurance",
                      "registration",
                      "other"
                    ],
                    "example": "inspection"
                  },
                  "description": {
                    "type": "string",
                    "example": "Documento de prueba para practicar subida con FormData"
                  },
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                },
                "required": [
                  "file"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Document stored successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadedCarDocumentResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid multipart payload or missing file"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Vehicle not found"
          },
          "413": {
            "description": "Uploaded file exceeds the 5 MB limit"
          },
          "415": {
            "description": "Unsupported file type"
          }
        },
        "security": [
          {
            "access_token": []
          }
        ],
        "summary": "Upload or replace the single stored document for a vehicle (ADMIN only)",
        "tags": [
          "Vehicles"
        ]
      }
    },
    "/cars/{id}/document/download": {
      "get": {
        "operationId": "CarsController_downloadCarDocument",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Vehicle identifier",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Document download started successfully"
          },
          "404": {
            "description": "Vehicle or document not found"
          }
        },
        "security": [
          {
            "access_token": []
          }
        ],
        "summary": "Download the single document linked to a vehicle",
        "tags": [
          "Vehicles"
        ]
      }
    },
...
  },
  "info": {
    "title": "NTT Demo Vehicle Catalog API",
    "description": "Backend API for the NTT Angular practice: authentication, vehicle catalog, brands, models, pagination and role-based access.\n\nRecommended flow in Swagger: call `POST /auth/login`, then use `GET /auth/me` and the protected endpoints normally. When the short-lived access cookie expires, call `POST /auth/refresh` and retry the failed request. Use `POST /auth/logout` to close the session cleanly.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "refresh_token"
      }
    },
    "schemas": {
...
      "PaginatedMetaDto": {
        "type": "object",
        "properties": {
          "totalItems": {
            "type": "number",
            "description": "Total number of records matching the current query",
            "example": 50
          },
          "itemCount": {
            "type": "number",
            "description": "Number of records returned in the current page",
            "example": 10
          },
          "itemsPerPage": {
            "type": "number",
            "description": "Maximum records per page requested by the client",
            "example": 10
          },
          "totalPages": {
            "type": "number",
            "description": "Total number of pages for the current query and limit",
            "example": 5
          },
          "currentPage": {
            "type": "number",
            "description": "Current page number (1-indexed)",
            "example": 1
          },
          "hasNextPage": {
            "type": "boolean",
            "description": "Indicates whether another page exists after the current one",
            "example": true
          },
          "hasPreviousPage": {
            "type": "boolean",
            "description": "Indicates whether a page exists before the current one",
            "example": false
          }
        },
        "required": [
          "totalItems",
          "itemCount",
          "itemsPerPage",
          "totalPages",
          "currentPage",
          "hasNextPage",
          "hasPreviousPage"
        ]
      },
      "PaginatedResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "description": "Collection of records returned for the current page",
            "type": "array",
            "items": {
              "type": "array"
            }
          },
          "meta": {
            "description": "Pagination metadata for the current query",
            "allOf": [
              {
                "$ref": "#/components/schemas/PaginatedMetaDto"
              }
            ]
          }
        },
        "required": [
          "items",
          "meta"
        ]
      },
      "CarBrandSummary": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Resolved brand identifier",
            "example": "brand-1"
          },
          "name": {
            "type": "string",
            "description": "Resolved brand display name",
            "example": "Toyota"
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
      "CarModelSummary": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Resolved model identifier",
            "example": "model-1"
          },
          "name": {
            "type": "string",
            "description": "Resolved model display name",
            "example": "Corolla"
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
      "CarSummary": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Vehicle identifier"
          },
          "brand": {
            "description": "Resolved brand information for read operations",
            "allOf": [
              {
                "$ref": "#/components/schemas/CarBrandSummary"
              }
            ]
          },
          "model": {
            "description": "Resolved model information for read operations",
            "allOf": [
              {
                "$ref": "#/components/schemas/CarModelSummary"
              }
            ]
          },
          "total": {
            "type": "number",
            "description": "Total number of vehicle detail entries"
          },
          "imageUrl": {
            "type": "string",
            "description": "Representative vehicle image URL for list and card views",
            "example": "/images/car_images/model-1_toyota_corolla.webp"
          }
        },
        "required": [
          "id",
          "brand",
          "model"
        ]
      },
...
      "Car": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Vehicle identifier"
          },
          "brand": {
            "description": "Resolved brand information for read operations",
            "allOf": [
              {
                "$ref": "#/components/schemas/CarBrandSummary"
              }
            ]
          },
          "model": {
            "description": "Resolved model information for read operations",
            "allOf": [
              {
                "$ref": "#/components/schemas/CarModelSummary"
              }
            ]
          },
          "carDetails": {
            "description": "Vehicle detail entries",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CarDetailEntity"
            }
          },
          "total": {
            "type": "number",
            "description": "Total number of vehicle detail entries"
          }
        },
        "required": [
          "id",
          "brand",
          "model"
        ]
      },
...
      "CreateCarDto": {
        "type": "object",
        "properties": {
          "brandId": {
            "type": "string",
            "description": "ID of the car brand. Must reference an existing brand (e.g. brand-1, brand-2).",
            "example": "brand-1"
          },
          "modelId": {
            "type": "string",
            "description": "ID of the car model. Must reference a model that belongs to the selected brand.",
            "example": "model-1"
          },
          "carDetails": {
            "description": "List of car detail entries (one per listing or unit). Can be omitted to create a car without details.",
            "example": [
              {
                "registrationDate": "2024-10-30T10:01:35.288Z",
                "mileage": 15000,
                "price": 20000,
                "manufactureYear": 2020,
                "currency": "EUR",
                "availability": true,
                "color": "Midnight Blue",
                "description": "Excellent condition, single owner.",
                "licensePlate": "1234 BBB"
              }
            ],
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CarDetailsDto"
            }
          }
        },
        "required": [
          "brandId",
          "modelId"
        ]
      },
      "UploadedCarDocumentResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Generated identifier for the uploaded document metadata",
            "example": "a3f7d1e3-5c7e-4a50-8b84-df73b46e5d4f"
          },
          "carId": {
            "type": "string",
            "description": "Vehicle identifier the upload was associated with",
            "example": "9f65ec7a-ef2c-4d8a-a7dc-248018fca712"
          },
          "originalName": {
            "type": "string",
            "description": "Original filename received in the multipart upload",
            "example": "itv.pdf"
          },
          "mimeType": {
            "type": "string",
            "description": "MIME type reported for the uploaded file",
            "example": "application/pdf"
          },
          "size": {
            "type": "number",
            "description": "File size in bytes",
            "example": 184532
          },
          "documentType": {
            "type": "string",
            "description": "Functional category assigned to the document",
            "enum": [
              "invoice",
              "inspection",
              "insurance",
              "registration",
              "other"
            ],
            "example": "inspection"
          },
          "title": {
            "type": "string",
            "description": "Optional display title for the document",
            "example": "Ficha tecnica ITV"
          },
          "description": {
            "type": "string",
            "description": "Optional notes attached to the document",
            "example": "Documento de prueba para practicar subida con FormData"
          },
          "uploadedAt": {
            "type": "string",
            "description": "Timestamp when the upload was received",
            "example": "2026-03-27T10:15:00.000Z"
          },
          "persisted": {
            "type": "boolean",
            "description": "Indicates whether the backend persisted the binary content",
            "example": true
          },
          "downloadUrl": {
            "type": "string",
            "description": "Download URL for the single document associated with the car",
            "example": "/cars/9f65ec7a-ef2c-4d8a-a7dc-248018fca712/document/download"
          },
          "message": {
            "type": "string",
            "description": "Explains how the backend handled the uploaded file",
            "example": "The file was stored on disk and replaced any previous document linked to the vehicle."
          }
        },
        "required": [
          "id",
          "carId",
          "originalName",
          "mimeType",
          "size",
          "documentType",
          "uploadedAt",
          "persisted",
          "downloadUrl",
          "message"
        ]
      },
...
    }
  }
}